% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_docsy.R
\name{document_class}
\alias{document_class}
\title{Document an R6 Class}
\usage{
document_class(aClass)
}
\arguments{
\item{aClass}{An \code{\link[R6]{R6Class}}. Note that this should be the class
generator, NOT an instance of that class.}
}
\description{
Given an R6 class, build a Roxygen documentation skeleton
             describing its public interface.
}
\examples{
\donttest{

# Build a sample class
LupeFiasco <- R6::R6Class(
    "LupeFiasco",
    public = list(
        initialize = function(food = TRUE, liquor = "interesting"){
            return(food & liquor)
        },
        kick_and_push = function(n = 10){
            return(sample(c("kick", "push", "coast"), size = n, replace = TRUE))
        },
        did_you_say_lupe_frasco = function(response_type = "lie"){
            switch(response_type,
                   "lie" = "yes",
                   "no")
        }
    ),
    active = list(
        they_call_me = function(){"lupe"},
        ill_be_your = function(){"new day"}
    )
)

# Document it!
cat(document_class(LupeFiasco))
}
}
