% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_package.R
\name{document_package}
\alias{document_package}
\title{Get R6 skeleton for a package}
\usage{
document_package(pkg_name, out_dir = file.path(getwd(), "roxy_docsy"))
}
\arguments{
\item{pkg_name}{Name of the package to document}

\item{out_dir}{Directory to write documentation to}
}
\description{
Given a package name, find al lthe R6 objects in that package's
             namespace and build skeleton Roxygen docs for each one. One file
             will be created in \code{out_dir} for each object.
}
