% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_R6_objects.R
\name{find_R6_objects}
\alias{find_R6_objects}
\title{Find R6 objects in a package}
\usage{
find_R6_objects(pkg_name)
}
\arguments{
\item{pkg_name}{Name of the package to document}
}
\description{
Search for R6 objects in a package and get a list of them
}
